[org.colinux.stable @ 0ef32fba018bff805a7e0859d7b8cc2e75e54313]  By Henry Nestler  Back ported from devel:  - SLiRP redirections, XML and command-line.  - Redesigned function slit_comma_seperated.    Fixes Param crash on BRIDGED_PCAP with too many params.  - WSAStartup before slirp_init.
git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/stable@304 688f6386-cf32-0410-9ead-a596f0484e40

+++ src/colinux/user/slirp/misc.c
@@ -819,7 +819,7 @@ fd_nonblock(fd)
 	int fd;
 {
 #ifdef FIONBIO
-	int opt = 1;
+	unsigned long opt = 1;
 	
 	ioctlsocket(fd, FIONBIO, &opt);
 #else
@@ -836,7 +836,7 @@ fd_block(fd)
 	int fd;
 {
 #ifdef FIONBIO
-	int opt = 0;
+	unsigned long opt = 0;
 	
 	ioctlsocket(fd, FIONBIO, &opt);
 #else
